% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dist_lookup_table}
\alias{dist_lookup_table}
\title{Lookup table for distribution abbreviations}
\format{
data.frame with two columns
\itemize{
  \item \code{dist} the abbreviated name of the distribution
  \item \code{name} the full name of the distribution
}
}
\usage{
data(dist_lookup_table)
}
\description{
A data frame with the full names and abbreviations for the distributions
in the `stats``
}
