% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qdist}
\alias{qdist}
\title{The general quantile function}
\usage{
qdist(dist, p, arg_list, lookup_verbose = F, ...)
}
\arguments{
\item{dist}{character string, specification of the distribution of interest using the abbreviations from \code{\link[stats]{distributions}}}

\item{p}{numeric vector of probabilities}

\item{arg_list}{list, alternative to directly using the names of the arguments for the distribution of choice, the user can provide a list with those arguments enclosed. This is useful for when multiple distributions need to be drawn from.}

\item{lookup_verbose}{logical, whether to print the abbreviated name found in the dist_lookup_table}

\item{...}{arguments specific to the distribution of interest (e.g. the "binom" distribution requires arguments for size and prob)}
}
\value{
a numeric vector representing the quantiles of the distribution of interest
}
\description{
The general quantile function
}
\examples{
## use curve and qdist to visualize the quantiles of a distribution
curve(qdist("pois", x, lambda=3), from=0, to=1)
}
