% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_eval}
\alias{safe_eval}
\title{Safer version of eval that only allows arthimetic operations}
\usage{
safe_eval(call)
}
\arguments{
\item{call}{an object to be evaluated}
}
\description{
Safer version of eval that only allows arthimetic operations
}
